Make box have default argument
